Added refresh on pull down on wallet page
